+++ tests/Other/Strategies/BlurryPercentileTest.php
@@ -11,15 +11,11 @@ class BlurryPercentileTest extends TestCase
 {
     protected $values;
 
-    protected $range;
-
     protected $strategy;
 
     public function setUp()
     {
         $this->values = [1, 2, 3, 4, 5];
-
-        $this->range = [0, 6];
 
         $this->strategy = new BlurryPercentile(50., 0.3);
     }
@@ -38,8 +34,8 @@ class BlurryPercentileTest extends TestCase
         $guess = $this->strategy->guess();
 
         $this->assertThat($guess, $this->logicalAnd(
-            $this->greaterThanOrEqual($this->range[0]),
-            $this->lessThanOrEqual($this->range[1]))
+            $this->greaterThanOrEqual(0),
+            $this->lessThanOrEqual(6))
         );
     }
 }