+++ tests/Other/Strategies/ConstantTest.php
@@ -2,37 +2,33 @@
 
 namespace Rubix\ML\Tests\Other\Strategies;
 
-use Rubix\ML\Other\Strategies\Mean;
+use Rubix\ML\Other\Strategies\Constant;
 use Rubix\ML\Other\Strategies\Strategy;
 use Rubix\ML\Other\Strategies\Continuous;
 use PHPUnit\Framework\TestCase;
 
-class MeanTest extends TestCase
+class ConstantTest extends TestCase
 {
-    protected $values;
-
     protected $strategy;
 
     public function setUp()
     {
-        $this->values = [1, 2, 3, 4, 5];
-
-        $this->strategy = new Mean();
+        $this->strategy = new Constant(17.);
     }
 
     public function test_build_strategy()
     {
-        $this->assertInstanceOf(Mean::class, $this->strategy);
+        $this->assertInstanceOf(Constant::class, $this->strategy);
         $this->assertInstanceOf(Continuous::class, $this->strategy);
         $this->assertInstanceOf(Strategy::class, $this->strategy);
     }
 
     public function test_make_guess()
     {
-        $this->strategy->fit($this->values);
+        $this->strategy->fit([]);
 
         $guess = $this->strategy->guess();
 
-        $this->assertEquals(3., $guess);
+        $this->assertEquals(17., $guess);
     }
 }