+++ CHANGELOG.md
@@ -7,6 +7,7 @@
     - Removed MetaEstimator interface
     - Added model Wrapper interface
     - AdaBoost is now probabilistic
+    - Added Constant guessing strategy
 
 - 0.0.7-alpha
     - Added Support Vector Machine classifier and regressor